+++ tcex/input/field_type/validator.py
@@ -18,6 +18,7 @@ def always_array(
     include_empty: bool = False,
     include_null: bool = False,
     split_csv: bool = False,
+    strip: bool = False,
 ) -> Callable[[Any, ModelField], list[Any]]:
     """Return customized validator that always returns a list.
 
@@ -29,6 +30,8 @@ def always_array(
             affect list with existing null values.
         split_csv: if True and input value is a string, then string will be split on comma. No
         further processing is done on result of splitting on comma
+        strip: if True and the input value is a string, then each string will be stripped of leading
+        and trailing whitespace.
     """
 
     def _always_array(value: Any, field: ModelField) -> list[Any]:
@@ -50,6 +53,9 @@ def always_array(
 
         if allow_empty is False and value == []:
             raise InvalidEmptyValue(field_name=field.name)
+
+        if strip is True:
+            value = [v.strip() if isinstance(v, str) else v for v in value]
 
         return value